Refactor is_current_run_pr_from_fork() to github_actions_api
Moved the implementation of is_current_run_pr_from_fork() from s3_buckets.py to github_actions_api.py for better modularity and reuse. Updated imports in compute_pytorch_cache_type.py to reflect this change, ensuring consistent access to the function across the codebase.
