+++ build_tools/_therock_utils/s3_buckets.py
@@ -142,43 +142,6 @@ def get_release_bucket_config(
     return _BUCKET_CONFIGS_BY_NAME[bucket_name]
 
 
-def _is_current_run_pr_from_fork() -> bool:
-    """Check if the current workflow run is a pull request from a fork.
-
-    Reads the GitHub event payload to check the .fork property on the
-    head repo, matching the behavior of the GitHub Actions expression
-    ``github.event.pull_request.head.repo.fork``.
-
-    Returns False for non-pull_request events or if the event payload
-    is not available (e.g. local development).
-    """
-    event_name = os.environ.get("GITHUB_EVENT_NAME", "")
-    if event_name != "pull_request":
-        return False
-
-    if not os.environ.get("GITHUB_EVENT_PATH"):
-        return False
-
-    # Deferred import: github_actions is optional in some environments; only
-    # needed when resolving fork state from the on-disk event payload.
-    from github_actions.github_actions_api import gha_load_github_event
-
-    event = gha_load_github_event()
-
-    return bool(
-        event.get("pull_request", {}).get("head", {}).get("repo", {}).get("fork", False)
-    )
-
-
-def is_current_run_pr_from_fork() -> bool:
-    """Whether the current workflow run is a pull request from a fork.
-
-    Public wrapper around the module-internal implementation, for reuse by
-    other build scripts.
-    """
-    return _is_current_run_pr_from_fork()
-
-
 def get_artifacts_bucket_config_for_workflow_run(
     github_repository: str,
     release_type: str | None = None,
@@ -240,7 +203,11 @@ def get_artifacts_bucket_config_for_workflow_run(
         _log(f"  head_github_repository: {head_github_repository}")
         _log(f"  is_pr_from_fork: {is_pr_from_fork}")
     else:
-        is_pr_from_fork = _is_current_run_pr_from_fork()
+        # Deferred import: github_actions is optional in some environments;
+        # only needed when resolving fork state from the on-disk event payload.
+        from github_actions.github_actions_api import is_current_run_pr_from_fork
+
+        is_pr_from_fork = is_current_run_pr_from_fork()
         _log(f"  is_pr_from_fork: {is_pr_from_fork}")
 
     config = get_artifacts_bucket_config(
