+++ build_tools/github_actions/github_actions_api.py
@@ -431,6 +431,26 @@ def gha_load_github_event() -> dict[str, Any]:
     return data
 
 
+def is_current_run_pr_from_fork() -> bool:
+    """Whether the current workflow run is a pull request from a fork.
+
+    Reads the GitHub event payload to check the ``.fork`` property on the head
+    repo, matching the GitHub Actions expression
+    ``github.event.pull_request.head.repo.fork``.
+
+    Returns False for non-pull_request events or if the event payload is not
+    available (e.g. local development).
+    """
+    if os.environ.get("GITHUB_EVENT_NAME", "") != "pull_request":
+        return False
+    if not os.environ.get("GITHUB_EVENT_PATH"):
+        return False
+    event = gha_load_github_event()
+    return bool(
+        event.get("pull_request", {}).get("head", {}).get("repo", {}).get("fork", False)
+    )
+
+
 def gha_send_request(url: str, timeout_seconds: int = 300) -> object:
     """Sends a request to the given GitHub REST API URL and returns the response.
